Legger til tilskudsprosent-tabell på godkjenningsiden
