+++ src/AvtaleSide/steg/BeregningTilskudd/BeregningVTAOTilskuddSteg.less
@@ -12,11 +12,6 @@
         margin-top: 1rem;
     }
 
-    &__tilskuddsprosent-heading,
-    &__tilskuddsprosent {
-        margin-bottom: 1.25rem;
-    }
-
     &__rad {
         margin-bottom: 2rem;
     }
