+++ src/AvtaleSide/steg/BeregningTilskudd/Lonnstilskuddprosent/index.ts
@@ -1,2 +1,4 @@
 import Lonnstilskuddprosent from './Lonnstilskuddprosent';
+import TilskuddstrinnTabell from './TilskuddstrinnTabell';
 export default Lonnstilskuddprosent;
+export { TilskuddstrinnTabell };
